wip: reorganize modules
