+++ src/application/create-application.ts
@@ -6,7 +6,7 @@ import { initializeRequestScopeStorage } from '@/Http/Scope/request-scope-storag
 import { registerEventSubscribers } from '@/Kernel';
 import { registerLegacyRoutes } from '@/routing/deprecated-decorator/legacy-router';
 import { verifyRoutingMode } from '@/routing/deprecated-decorator/verify-routing-mode';
-import { registerRoutes } from '@/routing/route/registration/register-routes';
+import { registerRoutes } from '@/routing/registration/register-routes';
 import type { Request } from 'koa';
 import Koa from 'koa';
 import { type Application } from './application';
