+++ src/index.ts
@@ -14,7 +14,7 @@ export * from '@/Kernel';
 // Routing
 export * from '@/routing/deprecated-decorator/route';
 export type { HttpMethod } from '@/routing/verb/http-method.type';
-export type { RouteOptions } from '@/routing/route/route-options';
+export type { RouteOptions } from '@/routing/route/route-options.type';
 export type { RouterMethod } from '@/routing/route/router-method.type';
 /**
  * @deprecated Use `Route` from `@koala-ts/framework/routing` instead.
