+++ src/routing/path/create-path-for.ts
@@ -1,6 +1,6 @@
-import { createPathCatalog } from '@/routing/route/source/create-path-catalog';
-import { createPathTemplateResolver } from '@/routing/route/source/resolve-path-template';
-import type { RouteSource } from '@/routing/route/source/route-source';
+import { createPathCatalog } from '@/routing/registration/source/create-path-catalog';
+import { createPathTemplateResolver } from '@/routing/registration/source/resolve-path-template';
+import type { RouteSource } from '@/routing/registration/source/route-source';
 
 type PathParamValue = string | number | boolean;
 type PathParams = Record<string, PathParamValue>;
