+++ src/routing/registration/source/create-path-catalog.test.ts
