+++ src/routing/registration/source/create-path-catalog.ts
@@ -1,5 +1,5 @@
-import { normalizeRouteSources } from '@/routing/route/source/normalize-route-sources';
-import type { RouteSource } from '@/routing/route/source/route-source';
+import { normalizeRouteSources } from '@/routing/registration/source/normalize-route-sources';
+import type { RouteSource } from '@/routing/registration/source/route-source';
 
 export function createPathCatalog(routeSources: RouteSource[]): Map<string, string> {
   const catalog = new Map<string, string>();
