+++ src/routing/registration/source/normalize-route-sources.test.ts
