+++ src/routing/registration/source/resolve-path-template.test.ts
