+++ src/routing/registration/source/resolve-path-template.ts
