+++ src/routing/registration/validation/validate-route-definitions.test.ts
